fixed setup problem in tests

+++ tests/test_dataservice.py
@@ -28,9 +28,11 @@ def test_data_service():
     
     pull_dir = Path(__file__).parent / "prep_data" / "test_full_flow"
     data_dir = Path(__file__).parent / "target"
-    for item in data_dir.glob("*"):
-        #print(f"removing {item}")
-        item.unlink()
+    if not data_dir.exists():
+        data_dir.mkdir()
+    else:
+        for item in data_dir.glob("*"):
+            item.unlink()
     for item in pull_dir.glob("*"):
         shutil.copy(item, data_dir)
     dataservice = DataService(data_dir)
